Safari-609.1.11.1/LayoutTests/http/tests/security/javascriptURL/xss-DENIED-to-javascript-url-in-foreign-domain-window-open-expected.txt

CONSOLE MESSAGE: line 45: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
PASS: Cross frame access to a javascript: URL embed in a frame window.open'ed on foreign domain denied!
